requirements.txt
line-bot-sdk==3.7.0
flask==3.0.0
azure-ai-translation-text==1.0.0
PyPDF2==3.0.1
requests==2.31.0
langchain==0.3.0
langchain-openai==0.2.0
python-dotenv==1.0.1
gunicorn==23.0.0
langchain-community==0.3.0
gunicorn>=22.0.0